refactor: (#123) 중복되는 기능 함수 합치기
